--- src/pages/Home.jsx
@@ -1,12 +1,0 @@
-import React from "react";
-import Header from "../components/Header";
-
-const Home = () => {
-  return (
-    <div>
-      <Header />
-    </div>
-  );
-};
-
-export default Home;
